Extract tool_policy_decisions.rs from turn.rs (parent #680)
Move the per-Agent tool-policy decision helpers out of `turn.rs` into
a focused sibling module so `turn.rs` keeps shrinking toward
responsibility-focused units (parent meta-issue #680).

Four production helpers extracted as free functions taking `&Agent`
(matching `actor_loop_flow` / `reply_retry` / earlier vertical-slice
precedent):

- `answer_only_mode_active` (Issue #576 / DR3-001 SSOT: trust the
  second-pass-corrected `session.mode_state.work_mode`)
- `script_execution_requested` (request explicitly requests local
  script execution)
- `answer_only_policy_error` (read-only gate with Read/Glob/Grep
  allowlist + Bash-script allowlist branch)
- `effective_tool_policy_error` (no-explicit-policy fallback via
  `effective_tool_policy_flow::effective_tool_policy`)

Behavior unchanged: same Issue #576 SSOT (no OR with lexical
pre-classifier), same Read/Glob/Grep / Bash-script allowlist order,
same workspace-scope override when a `MissingVerifierJob` is active.

`pub(super)` limited / no facade re-export (DR3-001). Callers updated
in `photon_feedback_derive.rs` (1 site),
`effective_tool_policy_flow.rs` (2 sites), `tool_call_execution.rs`
(2 sites), and `actor_loop_flow.rs` (2 sites) to the free-fn form.
Also gated `effective_tool_policy_error_for_call_with_scope` import in
`turn.rs` under `#[cfg(test)]` (the last non-test caller was the just-
extracted method; the remaining production-callable test seam
`drive_policy_error_for_test` still uses it directly).

LOC delta:
- turn.rs: 747 → 687 (-60)
- new module: +92

3,114 lib tests pass; `cargo fmt` / `cargo clippy --lib --all-targets
-- -D warnings` clean.
